Fix missing component imports
