+++ dgh-platform/chat-bot/frontend/lib/auth-context.tsx
@@ -1,17 +1,17 @@
+// File: lib/auth-context.tsx
 "use client"
 
-import type React from "react"
-import { createContext, useContext, useState } from "react"
+import React, { createContext, useContext, useState } from "react"
 
 interface User {
   id: string
-  name: string
   email: string
+  name: string
 }
 
 interface AuthContextType {
   user: User | null
-  login: (email: string, password: string) => Promise<void>
+  login: (email: string, password: string) => Promise<boolean>
   logout: () => void
   isLoading: boolean
 }
@@ -22,29 +22,49 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
   const [user, setUser] = useState<User | null>(null)
   const [isLoading, setIsLoading] = useState(false)
 
-  const login = async (email: string, password: string) => {
+  const login = async (email: string, password: string): Promise<boolean> => {
     setIsLoading(true)
+    
     // Simulate API call
-    await new Promise((resolve) => setTimeout(resolve, 1000))
-    setUser({
-      id: "1",
-      name: "John Doe",
-      email: email,
-    })
+    await new Promise(resolve => setTimeout(resolve, 1000))
+    
+    // Mock successful login
+    if (email && password) {
+      setUser({
+        id: "1",
+        email,
+        name: email.split("@")[0],
+      })
+      setIsLoading(false)
+      return true
+    }
+    
     setIsLoading(false)
+    return false
   }
 
   const logout = () => {
     setUser(null)
   }
 
-  return <AuthContext.Provider value={{ user, login, logout, isLoading }}>{children}</AuthContext.Provider>
+  const value = {
+    user,
+    login,
+    logout,
+    isLoading,
+  }
+
+  return (
+    <AuthContext.Provider value={value}>
+      {children}
+    </AuthContext.Provider>
+  )
 }
 
 export function useAuth() {
   const context = useContext(AuthContext)
-  if (!context) {
-    throw new Error("useAuth must be used within AuthProvider")
+  if (context === undefined) {
+    throw new Error("useAuth must be used within an AuthProvider")
   }
   return context
 }
