+++ dgh-platform/chat-bot/frontend/lib/files-context.tsx
@@ -107,124 +107,3 @@ export function useFiles(): FilesContextType {
   }
   return context
 }
-
-// File: lib/theme-context.tsx
-"use client"
-
-import React, { createContext, useContext, useEffect, useState } from "react"
-
-type Theme = "dark" | "light" | "system"
-
-interface ThemeContextType {
-  theme: Theme
-  setTheme: (theme: Theme) => void
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined)
-
-export function ThemeProvider({ children }: { children: React.ReactNode }) {
-  const [theme, setTheme] = useState<Theme>("system")
-
-  useEffect(() => {
-    const root = window.document.documentElement
-    const systemTheme = window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light"
-    
-    root.classList.remove("light", "dark")
-    
-    if (theme === "system") {
-      root.classList.add(systemTheme)
-    } else {
-      root.classList.add(theme)
-    }
-  }, [theme])
-
-  const value = {
-    theme,
-    setTheme,
-  }
-
-  return (
-    <ThemeContext.Provider value={value}>
-      {children}
-    </ThemeContext.Provider>
-  )
-}
-
-export function useTheme() {
-  const context = useContext(ThemeContext)
-  if (context === undefined) {
-    throw new Error("useTheme must be used within a ThemeProvider")
-  }
-  return context
-}
-
-// File: lib/auth-context.tsx
-"use client"
-
-import React, { createContext, useContext, useState } from "react"
-
-interface User {
-  id: string
-  email: string
-  name: string
-}
-
-interface AuthContextType {
-  user: User | null
-  login: (email: string, password: string) => Promise<boolean>
-  logout: () => void
-  isLoading: boolean
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined)
-
-export function AuthProvider({ children }: { children: React.ReactNode }) {
-  const [user, setUser] = useState<User | null>(null)
-  const [isLoading, setIsLoading] = useState(false)
-
-  const login = async (email: string, password: string): Promise<boolean> => {
-    setIsLoading(true)
-    
-    // Simulate API call
-    await new Promise(resolve => setTimeout(resolve, 1000))
-    
-    // Mock successful login
-    if (email && password) {
-      setUser({
-        id: "1",
-        email,
-        name: email.split("@")[0],
-      })
-      setIsLoading(false)
-      return true
-    }
-    
-    setIsLoading(false)
-    return false
-  }
-
-  const logout = () => {
-    setUser(null)
-  }
-
-  const value = {
-    user,
-    login,
-    logout,
-    isLoading,
-  }
-
-  return (
-    <AuthContext.Provider value={value}>
-      {children}
-    </AuthContext.Provider>
-  )
-}
-
-export function useAuth() {
-  const context = useContext(AuthContext)
-  if (context === undefined) {
-    throw new Error("useAuth must be used within an AuthProvider")
-  }
-  return context
-}
